+++ apps/indexer/lib/indexer/supervisor.ex
@@ -25,6 +25,7 @@ defmodule Indexer.Supervisor do
 
   alias Indexer.Temporary.{
     BlocksTransactionsMismatch,
+    InternalTransactionsBlockNumber,
     UncatalogedTokenTransfers,
     UnclesWithoutIndex
   }
@@ -127,7 +128,8 @@ defmodule Indexer.Supervisor do
         {UnclesWithoutIndex.Supervisor,
          [[json_rpc_named_arguments: json_rpc_named_arguments, memory_monitor: memory_monitor]]},
         {BlocksTransactionsMismatch.Supervisor,
-         [[json_rpc_named_arguments: json_rpc_named_arguments, memory_monitor: memory_monitor]]}
+         [[json_rpc_named_arguments: json_rpc_named_arguments, memory_monitor: memory_monitor]]},
+        {InternalTransactionsBlockNumber.Supervisor, [[memory_monitor: memory_monitor]]}
       ],
       strategy: :one_for_one
     )